==> 2_feature_engineering.Rmd <==
---
title: "Feature Engineering"
output: html_notebook
---

Summary: In this notebook, I engineer two features to help with covid 19 predictions:

- days since first case
- days since first fatility

```{r}
library(data.table)
library(tidyverse); theme_set(theme_bw(base_size = 20))
library(prophet)
library(ggpubr)
library(ggsci)
library(SuperLearner)
```

```{r}
covid.a <- read_csv("~/Documents/data/ph244/data-proj-4a.csv")
```

```{r}
first_confirmed_case <- covid.a %>% 
  filter(ConfirmedCases > 0) %>% 
  group_by(Country_Region, Province_State) %>% 
  summarize(first_confirmed_case = min(Date))

first_confirmed_fatility <- covid.a %>% 
  filter(Fatalities > 0) %>% 
  group_by(Country_Region, Province_State) %>% 
  summarize(first_confirmed_fatility = min(Date))


covid.a <- covid.a %>% 
  left_join(first_confirmed_case) %>% 
  left_join(first_confirmed_fatility)
```


```{r}
covid.a <-  covid.a %>% 
  mutate(days_since_first_case =  as.numeric(Date - first_confirmed_case)) %>% 
  mutate(days_since_first_fatility =  as.numeric(Date - first_confirmed_fatility)) %>% 
  mutate(geo = paste(Country_Region, Province_State, sep = "_")) %>% 
  mutate(days_since_first_case = case_when(
    days_since_first_case < 0 |is.na(days_since_first_fatility) ~ 0, 
    TRUE ~ days_since_first_case
  )) %>% 
  mutate(days_since_first_fatility = case_when(
    days_since_first_fatility < 0 | is.na(days_since_first_fatility) ~ 0, 
    TRUE ~ days_since_first_fatility
  )) %>% 
  mutate(training = case_when(
    Date < "2020-03-18" ~ 1,
    Date >= "2020-03-18" ~ 0
  ))

write_csv(covid.a, "~/Documents/data/ph244/intermediate/test_training_split.csv")
```
